Refactor sidenav out of app top level
